pyproject.toml
[project]
name = "piot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "influxdb3-python>=0.16.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "uvicorn>=0.37.0",
]
